[examples,conf] draft wham recipe
