+++ conf/wham/1b.yaml
@@ -1,0 +1,47 @@
+# Conv-TasNet
+
+nnet: sse@time_tasnet
+
+nnet_conf:
+  L: 40
+  N: 256
+  X: 8
+  R: 4
+  B: 256
+  H: 512
+  P: 3
+  norm: "cLN"
+  num_spks: 2
+  non_linear: "relu"
+
+task: "sse@sisnr"
+
+task_conf:
+  num_spks: 2
+  permute: true
+  zero_mean: false
+
+trainer_conf:
+  optimizer: "adam"
+  optimizer_kwargs:
+    lr: 1.0e-3
+    weight_decay: 1.0e-5
+  lr_scheduler_kwargs:
+    min_lr: 1.0e-8
+    patience: 1
+    factor: 0.5
+  clip_gradient: 10
+  no_impr: 6
+  no_impr_thres: 0.1
+
+data_conf:
+  fmt: "se@chunk"
+  loader:
+    chunk_size: 64000
+    sr: 16000
+  train:
+    mix_scp: "data/wham_16k_min/tr/mix.scp"
+    ref_scp: "data/wham_16k_min/tr/spk1.scp,data/wham_16k_min/tr/spk2.scp"
+  valid:
+    mix_scp: "data/wham_16k_min/cv/mix.scp"
+    ref_scp: "data/wham_16k_min/cv/spk1.scp,data/wham_16k_min/cv/spk2.scp"
