+++ examples/sse/wham/cmd
@@ -1,0 +1,1 @@
+../../../cmd
