+++ examples/sse/wham/conf
@@ -1,0 +1,1 @@
+../../../conf
