+++ examples/sse/wham/scripts
@@ -1,0 +1,1 @@
+../../../scripts
